Atualização Homolog Sprint 5

+++ sme_ptrf_apps/core/models/relacao_bens.py
@@ -49,7 +49,6 @@ def auto_delete_file_on_change(sender, instance, **kwargs):
         return False
 
     new_file = instance.arquivo
-    if not old_file == new_file:
+    if old_file and not old_file == new_file:
         if os.path.isfile(old_file.path):
             os.remove(old_file.path)
-
